+++ src/lib/microcms/index.ts
@@ -1,0 +1,1 @@
+export { client } from "./client";
